充值、提现、优惠券、收藏、实名认证、更换社区、常见问题、违规、举报，服务获取优惠券、领取优惠券

+++ app/src/main/java/com/lixin/amuseadjacent/app/ui/mine/request/UserInfo_19.kt
@@ -22,7 +22,7 @@ object UserInfo_19 {
                 super.onResponse(response, id)
                 val model = Gson().fromJson(response, UserInfoModel::class.java)
                 if (model.result == "0") {
-                    StaticUtil.balance = model.balance.toDouble()
+                    StaticUtil.balance = model.balance
                     StaticUtil.nickName = model.nickname
                     StaticUtil.effectNum = model.effectNum
                     EventBus.getDefault().post(model)
